Update README.md
A machine learning project which has 2 parts. In the first part, face recognition is carried out, where model is trained using an input dataset and tested by giving a random input image and the models recognizes the person and also gives its similarity to other faces in the dataset.Second part - Mask detection. Model detects usinh the live cam and tells if a person has worn a mask or not.